refs #7: revamp option (de)serialization
There are now two strategies:

 * OptionFormatAsSelf, the previous strategy which encodes to self/null
 * OptionFormatAsArray, which encodes Some(x) / None to [x] / []

By using LowLevelImplicits we provide sensible defaults for nested options
which is to use the array encoding for outer options in a nested setting
and the self/null-encoding for the innermost option or simple options.

The leaf-level option format can be customized by bringing an instance of
LeafLevelOptionFormat into scope.