spray-json: in monadic operations require explicit conversions
This is needed because with Scala 2.10 an expression

    for (name: String <- json)

is rewritten to

    json.withFilter {
      case x: String => true
      case _ => false
    }

Because there is no way to define an implicit conversion
which would automatically lift the PartialFunction to one with automatic
JsonReader support, we now have to rely on explicit conversions.

The new way to write the above expression is

    for (name <- json.toValidated[String])

which is only longer because `toValidated` is currently too verbose.