! http: make only standard charsets available as constants, fixes #340
Only standard charsets are known to be available on all Java platforms.
If another one was missing, initialization fatally aborted. Constants
for others charsets can still be retrieved from the registry but can't
be supplied as constants as those may not be available.

Another bug was fixed here: the 'ISO-8859-10' charset was wrongly registered
as 'ISO-8859-1'.