! routing: Added `detach` directive which executes its inner route in a future. Removed `detachTo` directive. Fixes #240.
The `detachTo` directive was removed and replaced with `detach` which executes its inner route in a `Future` (as opposed to
executing it inside an actor). In order to migrate, users will have to replace their use of `detachTo`, which most commonly
was used with `singleRequestServiceActor`, with a parameterless call to detach (i.e. `detach()`) if they have either an
implicit `ExecutionContext` or `ActorRefFactory` in scope. Additionally, it is also possible to call detach with an explicit `ExecutionContext` (e.g. `detach(ec)`).