= httpx: unwrap `MalformedContent` error in `ResponseTransformation`
When unmarshalling in the pipeline, MalformedContent exceptions were
converted to strings and set as the message for a PipelineException.
However, this suppresses the stacktrace of the underlying cause (if
there is one), making debugging a custom/delegate unmarshaller very
difficult.

This commit sets the cause of the PipelineException to that of the
MalformedContent exception (or null if it does not exist).