+ http: add support for RFC 5861 stale content Cache-Control extensions
Support for RFC 5861 'HTTP Cache-Control Extensions for Stale Content'.

The stale-while-revalidate Cache-Control Extension

When present in an HTTP response, the stale-while-revalidate
Cache-Control extension indicates that caches MAY serve the response in
which it appears after it becomes stale, up to the indicated number of
seconds.

http://tools.ietf.org/html/rfc5861#section-3

The stale-if-error Cache-Control Extension

When present in an HTTP response, this value indicates the upper limit
to staleness; when the cached response is more stale than the indicated
the cached response SHOULD NOT be used to satisfy the request.

http://tools.ietf.org/html/rfc5861#section-4